% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adbc.R
\name{adbc_connection_init}
\alias{adbc_connection_init}
\alias{adbc_connection_init_default}
\alias{adbc_connection_set_options}
\alias{adbc_connection_release}
\title{Connections}
\usage{
adbc_connection_init(database, ...)

adbc_connection_init_default(database, options = NULL, subclass = character())

adbc_connection_set_options(connection, options)

adbc_connection_release(connection)
}
\arguments{
\item{database}{An \link[=adbc_database_init]{adbc_database}.}

\item{...}{Driver-specific options. For the default method, these are
named values that are converted to strings.}

\item{options}{A named \code{character()} or \code{list()} whose values are converted
to strings.}

\item{subclass}{An extended class for an object so that drivers can specify
finer-grained control over behaviour at the R level.}

\item{connection}{An \link[=adbc_connection_init]{adbc_connection}}
}
\value{
An object of class 'adbc_connection'
}
\description{
Connections
}
\examples{
db <- adbc_database_init(adbc_driver_void())
adbc_connection_init(db)

}
