% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver_monkey.R
\name{adbc_driver_monkey}
\alias{adbc_driver_monkey}
\title{Monkey see, monkey do!}
\usage{
adbc_driver_monkey()
}
\value{
An object of class 'adbc_driver_monkey'
}
\description{
A driver whose query results are set in advance.
}
\examples{
db <- adbc_database_init(adbc_driver_monkey())
con <- adbc_connection_init(db)
stmt <- adbc_statement_init(con, mtcars)
stream <- nanoarrow::nanoarrow_allocate_array_stream()
adbc_statement_execute_query(stmt, stream)
as.data.frame(stream$get_next())

}
