% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver_void.R
\name{adbc_driver_void}
\alias{adbc_driver_void}
\alias{adbc_driver}
\alias{adbc_driver_load_raw}
\alias{adbc_driver_resolve}
\alias{adbc_driver_load}
\title{Create ADBC drivers}
\usage{
adbc_driver_void()

adbc_driver(x, entrypoint = NULL, ..., subclass = character())

adbc_driver_load_raw(x, entrypoint, version, driver, error)

adbc_driver_resolve(x, entrypoint)

adbc_driver_load(x, entrypoint, version, driver, error)
}
\arguments{
\item{x, entrypoint}{An ADBC driver may be defined either as an init function
or as an identifier with an entrypoint name. A driver init func
must be an external pointer to a DL_FUNC with the type
\code{AdbcDriverInitFunc} specified in the adbc.h header.}

\item{...}{Further key/value parameters to store with the (R-level) driver
object.}

\item{subclass}{An optional subclass for finer-grained control of
behaviour at the R level.}

\item{version}{The version number corresponding to the \code{driver} supplied}

\item{driver}{An external pointer to an \code{AdbcDriver}}

\item{error}{An external pointer to an \code{AdbcError} or NULL}
}
\value{
\itemize{
\item \code{adbc_driver()}, \code{adbc_driver_void()}: An object of class 'adbc_driver'
\item \code{adbc_driver_load_raw()}: An integer status code.
}
}
\description{
Creates the R object representation of an ADBC driver, which consists of a
name and an initializer function with an optional subclass to control
finer-grained behaviour at the R level.
}
\examples{
adbc_driver_void()

}
