% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-dbplyr.R
\name{adbc_simulate_dbi}
\alias{adbc_simulate_dbi}
\title{Simulate an equivalent DBI connection}
\usage{
adbc_simulate_dbi(connection, method = NULL, ...)
}
\arguments{
\item{connection}{An \link[=adbc_connection_init]{adbc_connection}}

\item{method}{The package-qualified name of the method that
requires a simulation of a DBI object.}

\item{...}{Passed to S3 methods}
}
\value{
An S3 object like that returned by \code{\link[dbplyr:simulate_dbi]{dbplyr::simulate_dbi()}}
}
\description{
Simulate an equivalent DBI connection
}
\examples{
\dontshow{if (requireNamespace("dbplyr", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
db <- adbc_database_init(adbc_driver_void())
con <- adbc_connection_init(db)
adbc_simulate_dbi(con)
\dontshow{\}) # examplesIf}
}
