% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adbc.R, R/options.R
\name{adbc_statement_init}
\alias{adbc_statement_init}
\alias{adbc_statement_init_default}
\alias{adbc_statement_release}
\alias{adbc_statement_set_options}
\alias{adbc_statement_get_option}
\alias{adbc_statement_get_option_bytes}
\alias{adbc_statement_get_option_int}
\alias{adbc_statement_get_option_double}
\title{Statements}
\usage{
adbc_statement_init(connection, ...)

adbc_statement_init_default(connection, options = NULL, subclass = character())

adbc_statement_release(statement)

adbc_statement_set_options(statement, options)

adbc_statement_get_option(statement, option)

adbc_statement_get_option_bytes(statement, option)

adbc_statement_get_option_int(statement, option)

adbc_statement_get_option_double(statement, option)
}
\arguments{
\item{connection}{An \link[=adbc_connection_init]{adbc_connection}}

\item{...}{Driver-specific options. For the default method, these are
named values that are converted to strings.}

\item{options}{A named \code{character()} or \code{list()} whose values are converted
to strings.}

\item{subclass}{An extended class for an object so that drivers can specify
finer-grained control over behaviour at the R level.}

\item{statement}{An \link[=adbc_statement_init]{adbc_statement}}

\item{option}{A specific option name}
}
\value{
An object of class 'adbc_statement'
}
\description{
Statements
}
\examples{
db <- adbc_database_init(adbc_driver_void())
con <- adbc_connection_init(db)
adbc_statement_init(con)

}
