% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-compat.R
\name{adbc_test_s4_compat}
\alias{adbc_test_s4_compat}
\title{S4 Compatibility}
\usage{
adbc_test_s4_compat(driver, ...)
}
\arguments{
\item{driver}{An \code{\link[=adbc_driver]{adbc_driver()}}}

\item{...}{Database parameters passed to \code{\link[=adbc_database_init]{adbc_database_init()}}}
}
\value{
A S4 object of class "AdbcS4CompatTest" with slots "driver",
"database", "connection", and "statement".
}
\description{
While wrappers around ADBC objects are implemented as S3 objects, these
wrappers should also be compatible with the S4 object system. This ensures
compatibility with S4 frameworks like DBI that define specific types for
slot values.
}
\examples{
adbc_test_s4_compat(adbc_driver_void())

}
