% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adbcdrivermanager-package.R
\docType{package}
\name{adbcdrivermanager-package}
\alias{adbcdrivermanager}
\alias{adbcdrivermanager-package}
\title{adbcdrivermanager: 'Arrow' Database Connectivity ('ADBC') Driver Manager}
\description{
Provides a developer-facing interface to 'Arrow' Database Connectivity ('ADBC') for the purposes of driver development, driver testing, and building high-level database interfaces for users.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paleolimbot/adbcdrivermanager}
  \item Report bugs at \url{https://github.com/paleolimbot/adbcdrivermanager/issues}
}

}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@dunnington.ca} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})

}
\keyword{internal}
