% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adbc.R
\name{adbc_connection_get_info}
\alias{adbc_connection_get_info}
\alias{adbc_connection_get_objects}
\alias{adbc_connection_get_table_schema}
\alias{adbc_connection_get_table_types}
\alias{adbc_connection_read_partition}
\alias{adbc_connection_commit}
\alias{adbc_connection_rollback}
\title{Connection methods}
\usage{
adbc_connection_get_info(connection, info_codes)

adbc_connection_get_objects(
  connection,
  depth,
  catalog,
  db_schema,
  table_name,
  table_type,
  column_name
)

adbc_connection_get_table_schema(connection, catalog, db_schema, table_name)

adbc_connection_get_table_types(connection)

adbc_connection_read_partition(connection, serialized_partition)

adbc_connection_commit(connection)

adbc_connection_rollback(connection)
}
\arguments{
\item{connection}{An \link[=adbc_connection_init]{adbc_connection}}

\item{info_codes}{(Currently undocumented)}

\item{depth}{(Currently undocumented)}

\item{catalog}{(Currently undocumented)}

\item{db_schema}{(Currently undocumented)}

\item{table_name}{(Currently undocumented)}

\item{table_type}{(Currently undocumented)}

\item{column_name}{(Currently undocumented)}

\item{serialized_partition}{(Currently undocumented)}
}
\value{
\itemize{
\item \code{adbc_connection_get_info()}, \code{adbc_connection_get_objects},
\code{adbc_connection_get_table_types()}, and \code{adbc_connection_read_partition()}
return a \link[nanoarrow:as_nanoarrow_array_stream]{nanoarrow_array_stream}.
\item \code{adbc_connection_get_table_schema()} returns a
\link[nanoarrow:as_nanoarrow_schema]{nanoarrow_schena}
\item \code{adbc_connection_commit()} and \code{adbc_connection_rollback()} return
\code{connection}, invisibly.
}
}
\description{
Connection methods
}
\examples{
db <- adbc_database_init(adbc_driver_void())
con <- adbc_connection_init(db)
# (not implemented by the void driver)
try(adbc_connection_get_info(con, 0))

}
