% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-dbi.R
\name{adbc_simulate_dbi}
\alias{adbc_simulate_dbi}
\alias{adbc_simulate_dbi_default}
\title{Simulate an equivalent DBI connection}
\usage{
adbc_simulate_dbi(connection, method = NULL, ...)

adbc_simulate_dbi_default(connection, class = character())
}
\arguments{
\item{connection}{An \link[=adbc_connection_init]{adbc_connection}}

\item{method}{The package-qualified name of the method that
requires a simulation of a DBI object.}

\item{...}{Passed to S3 methods}

\item{class}{The DBIConnection subclass for the simulated
connection.}
}
\value{
An S3 object like that returned by \code{dbplyr::simulate_dbi()}
}
\description{
Simulate an equivalent DBI connection
}
\examples{
db <- adbc_database_init(adbc_driver_void())
con <- adbc_connection_init(db)
adbc_simulate_dbi(con)

}
