% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adbcdrivermanager-package.R
\docType{package}
\name{adbcdrivermanager-package}
\alias{adbcdrivermanager}
\alias{adbcdrivermanager-package}
\title{adbcdrivermanager: 'Arrow' Database Connectivity ('ADBC') Driver Manager}
\description{
Provides a developer-facing interface to 'Arrow' Database Connectivity ('ADBC') for the purposes of driver development, driver testing, and building high-level database interfaces for users. 'ADBC' \url{https://arrow.apache.org/adbc/} is an API standard for database access libraries that uses 'Arrow' for result sets and query parameters.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/apache/arrow-adbc}
  \item \url{https://arrow.apache.org/adbc/current/r/adbcdrivermanager/}
  \item Report bugs at \url{https://github.com/apache/arrow-adbc/issues}
}

}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@dunnington.ca} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})

Authors:
\itemize{
  \item Apache Arrow \email{dev@arrow.apache.org} [copyright holder]
}

Other contributors:
\itemize{
  \item Apache Software Foundation \email{dev@arrow.apache.org} [copyright holder]
}

}
\keyword{internal}
