% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{with_adbc_database}
\alias{with_adbc_database}
\alias{with_adbc_connection}
\alias{with_adbc_statement}
\alias{local_adbc_database}
\alias{local_adbc_connection}
\alias{local_adbc_statement}
\title{Cleanup helpers}
\usage{
with_adbc_database(database, code)

with_adbc_connection(connection, code)

with_adbc_statement(statement, code)

local_adbc_database(database, .local_envir = parent.frame())

local_adbc_connection(connection, .local_envir = parent.frame())

local_adbc_statement(statement, .local_envir = parent.frame())
}
\arguments{
\item{database}{A database created with \code{\link[=adbc_database_init]{adbc_database_init()}}}

\item{code}{Code to execute before cleaning up the input.}

\item{connection}{A connection created with \code{\link[=adbc_connection_init]{adbc_connection_init()}}}

\item{statement}{A statement created with \code{\link[=adbc_statement_init]{adbc_statement_init()}}}

\item{.local_envir}{The execution environment whose scope should be tied
to the input.}
}
\value{
\itemize{
\item \verb{with_*()} variants return the result of \code{code}
\item \verb{local_*()} variants return the input, invisibly.
}
}
\description{
Managing the lifecycle of databases, connections, and statements can
be complex and error-prone. The R objects that wrap the underlying ADBC
pointers will perform cleanup in the correct order if you rely on garbage
collection (i.e., do nothing and let the objects go out of scope); however
it is good practice to explicitly clean up these objects. These helpers
are designed to make explicit and predictable cleanup easy to accomplish.
}
\details{
Note that you can use \code{\link[=adbc_connection_join_database]{adbc_connection_join_database()}},
\code{\link[=adbc_statement_join_connection]{adbc_statement_join_connection()}}, and \code{\link[=adbc_stream_join_statement]{adbc_stream_join_statement()}}
to tie the lifecycle of the parent object to that of the child object.
These functions mark any previous references to the parent object as
released so you can still use local and with helpers to manage the parent
object before it is joined.
}
\examples{
# Using with_adbc_*():
with_adbc_database(db <- adbc_database_init(adbc_driver_void()), {
  with_adbc_connection(con <- adbc_connection_init(db), {
    with_adbc_statement(stmt <- adbc_statement_init(con), {
      # adbc_statement_set_sql_query(stmt, "SELECT * FROM foofy")
      # adbc_statement_execute_query(stmt)
    })
  })
})

# Using local_adbc_*() (works best within a function, test, or local())
local({
  db <- local_adbc_database(adbc_database_init(adbc_driver_void()))
  con <- local_adbc_connection(adbc_connection_init(db))
  stmt <- local_adbc_statement(adbc_statement_init(con))
  # adbc_statement_set_sql_query(stmt, "SELECT * FROM foofy")
  # adbc_statement_execute_query(stmt)
})

}
